use built in testrunner
