+++ gather2-core/gather2/templates/rest_framework/api.html
