+++ gather2-core/gather2/templates/surveyitemviewset.html
