+++ gather2-core/gather2/templates/surveyviewset_list.html
