+++ gather2-core/gather2/urls.py
@@ -15,11 +15,10 @@ Including another URLconf
 """
 from django.conf.urls import url, include
 
-from . import views
+from core import views
 from .routers import TemplateRouter
 from django.contrib import admin
 
-import gather2_odk_import.urls
 
 router = TemplateRouter(template_name='index.html')
 (
@@ -36,6 +35,4 @@ urlpatterns = [
     url(r'^api-auth/', include('rest_framework.urls',
                                namespace='rest_framework')),
     url(r'^admin/', include(admin.site.urls)),
-    url(r'^odk/', include(gather2_odk_import.urls)),
-
 ]
