+++ gather2-core/gather2/wsgi.py
@@ -9,7 +9,7 @@ https://docs.djangoproject.com/en/dev/howto/deployment/wsgi/
 
 import os
 
-from configurations.wsgi import get_wsgi_application
+from django.core.wsgi import get_wsgi_application
 
 os.environ.setdefault("DJANGO_SETTINGS_MODULE", "gather2.settings")
