+++ gather2-core/requirements.txt
@@ -1,4 +1,7 @@
 Django>=1.9a1
+flake8==2.4.1
+autopep8==1.2.1
+ipdb==0.8.1
 djangorestframework==3.2
 drf-extensions==0.2.8
 psycopg2==2.6.1
@@ -6,5 +9,3 @@ django-filter==0.11.0
 dj_database_url==0.3.0
 jsonschema==2.5.1
 https://github.com/django-extensions/django-extensions/zipball/92008a391c6357d7896a65ea74ac103664e21ddc
-https://github.com/jezdez/django-configurations/zipball/0ff4c7612d28036449edd4b2cf3b58ca1eded0f2
-
