--- pytest.ini
@@ -1,5 +1,0 @@
-[pytest]
-DJANGO_SETTINGS_MODULE=core.settings
-DJANGO_CONFIGURATION=Test
-django_find_project = false
-
