--- setup.py
@@ -1,10 +1,0 @@
-from setuptools import setup, find_packages
-
-setup(
-    name="gather2",
-    version="0.0.0",
-    long_description=__doc__,
-    packages=find_packages(),
-    include_package_data=True,
-    zip_safe=False,
-)
